Remove generated block code, reorganize parser

+++ src/main/java/it/smartio/fastcc/jjtree/ASTProduction.java
@@ -50,7 +50,7 @@ public class ASTProduction extends JJTreeNode {
       i = this.nextNodeScopeNumber++;
       this.scopes.put(s, i);
     }
-    return i.intValue();
+    return i;
   }
 
   @Override
